CSL: OK so I updated to the latest macOS and with it the latest Xcode. The result was that some parts of the CSL Reduce would no longer compile and there were a LOT of new "this function is deprecated" messages that rendered compilation noisy and unsatisfactory. This checkin is in part so I can let people know I am working on these issues - though as of right now I can build csl.app (but woth more deprecation warnings than is at all nice), but bootstrapreduce.img fails to build in a way that I need to investigate and fix.
git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@6428 2bfe0521-f11c-4a00-b80e-6202646ff360

+++ csl/cslbase/arith.h
@@ -549,7 +549,7 @@ inline LispObject make_lisp_unsigned64(uint64_t n)
 // while I get confused others could too... so safety is good.
 
 inline LispObject make_lisp_integerptr(intptr_t n)
-{   if (valid_as_fixnum(static_cast<intptr_t>(n))) return fixnum_of_int(n);
+{   if (valid_as_fixnum(static_cast<int64_t>(n))) return fixnum_of_int(n);
     else return make_lisp_integer64_fn(static_cast<intptr_t>(n));
 }
